refactor: rename industry_type.py file name
