+++ backend/src/backend/engine/types/industry_type.py
