+++ backend/tests/engine/agents/conftest.py
@@ -1,6 +1,6 @@
 from mesa import Model
 import pytest
-from backend.engine.types.IndustryType import IndustryType
+from backend.engine.types.industry_type import IndustryType
 
 
 class MockEconomyModel(Model):
